src/requirements.txt
flask>=2.0.0
google-cloud-secret-manager>=2.0.0
google-cloud-logging>=3.0.0
requests>=2.25.0
functions-framework
